pyproject.toml
[tool.poetry]
name = "nuclei"
version = "0.3.2"
description = ""
authors = ["Wizock <silent.death3500@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
ffmpeg="1.4"
fastapi="0.79.0"
ffmpeg_python="0.2.0"
gevent="21.12.0"
httpx="0.23.0"
jose="1.0.0"
lz4="4.0.1"
passlib="1.7.4"
Pillow="9.2.0"
pydantic="1.9.1"
python_jose="3.3.0"
SQLAlchemy="1.4.35"
PyJWT="1.4.0"
tqdm="4.64.0"
typing_extensions="4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
